apps/backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
context==0.1.0
cryptography==46.0.3
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
Faker==37.11.0
fastapi==0.133.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.13.0
fastapi-mail==1.4.1
fastar==0.8.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
limits==5.6.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.53.0
shellingham==1.5.4
slowapi==0.1.9
SQLAlchemy==2.0.47
sqlmodel==0.0.37
starlette==0.52.1
tenacity==9.1.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
win32_setctime==1.2.0
wrapt==1.17.3
